Reworking ModelItem to allow access via String (as well as Index). Allow looking up child name from parent Allows implementing sorting via QTreeWidget sort function. Removed need to keep ModelItems in order with Indices

+++ Tests/SampleApp/SampleModel.h
@@ -26,7 +26,7 @@ public:
 		return m_children.size();
 	}
 
-	virtual BaseModelItem* GetChild(size_t i) /*override*/
+	virtual BaseModelItem* GetChild(int i) /*override*/
 	{
 		return m_children[i];
 	}
